+++ internal/otelcollector/config/gatewayprocs/resolve_service_name_proc.go
@@ -1,0 +1,48 @@
+package gatewayprocs
+
+import (
+	"fmt"
+	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config"
+)
+
+func MakeResolveServiceNameStatements() []config.TransformProcessorStatements {
+	attributes := []string{
+		"kyma.kubernetes_io_app_name",
+		"kyma.app_name",
+		"k8s.deployment.name",
+		"k8s.daemonset.name",
+		"k8s.statefulset.name",
+		"k8s.job.name",
+		"k8s.pod.name",
+	}
+
+	var statements []string
+	for _, attr := range attributes {
+		statements = append(statements, inferServiceNameFromAttr(attr))
+	}
+	statements = append(statements, setDefaultServiceName())
+
+	return []config.TransformProcessorStatements{
+		{
+			Context:    "resource",
+			Statements: statements,
+		},
+	}
+}
+
+const serviceNameNotDefinedCondition = "attributes[\"service.name\"] == nil or attributes[\"service.name\"] == \"\" or attributes[\"service.name\"] == \"unknown_service\""
+
+func inferServiceNameFromAttr(attrKey string) string {
+	return fmt.Sprintf(
+		"set(attributes[\"service.name\"], attributes[\"%s\"]) where %s",
+		attrKey,
+		serviceNameNotDefinedCondition,
+	)
+}
+
+func setDefaultServiceName() string {
+	return fmt.Sprintf(
+		"set(attributes[\"service.name\"], \"unknown_service\") where %s",
+		serviceNameNotDefinedCondition,
+	)
+}
