+++ internal/otelcollector/config/metric/gateway/processors.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 
 	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config"
+	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config/gatewayprocs"
 	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config/metric"
-	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config/servicename"
 )
 
 func makeProcessorsConfig() Processors {
@@ -13,7 +13,7 @@ func makeProcessorsConfig() Processors {
 		BaseProcessors: config.BaseProcessors{
 			Batch:         makeBatchProcessorConfig(),
 			MemoryLimiter: makeMemoryLimiterConfig(),
-			K8sAttributes: makeK8sAttributesProcessorConfig(),
+			K8sAttributes: gatewayprocs.MakeK8sAttributesProcessorConfig(),
 			Resource:      makeResourceProcessorConfig(),
 		},
 		CumulativeToDelta:  &CumulativeToDeltaProcessor{},
@@ -34,41 +34,6 @@ func makeMemoryLimiterConfig() *config.MemoryLimiter {
 		CheckInterval:        "0.1s",
 		LimitPercentage:      75,
 		SpikeLimitPercentage: 10,
-	}
-}
-
-func makeK8sAttributesProcessorConfig() *config.K8sAttributesProcessor {
-	k8sAttributes := []string{
-		"k8s.pod.name",
-		"k8s.node.name",
-		"k8s.namespace.name",
-		"k8s.deployment.name",
-		"k8s.statefulset.name",
-		"k8s.daemonset.name",
-		"k8s.cronjob.name",
-		"k8s.job.name",
-	}
-
-	podAssociations := []config.PodAssociations{
-		{
-			Sources: []config.PodAssociation{{From: "resource_attribute", Name: "k8s.pod.ip"}},
-		},
-		{
-			Sources: []config.PodAssociation{{From: "resource_attribute", Name: "k8s.pod.uid"}},
-		},
-		{
-			Sources: []config.PodAssociation{{From: "connection"}},
-		},
-	}
-
-	return &config.K8sAttributesProcessor{
-		AuthType:    "serviceAccount",
-		Passthrough: false,
-		Extract: config.ExtractK8sMetadata{
-			Metadata: k8sAttributes,
-			Labels:   servicename.ExtractLabels(),
-		},
-		PodAssociation: podAssociations,
 	}
 }
 
@@ -116,21 +81,7 @@ func makeDropIfInputSourceIstioConfig() *FilterProcessor {
 
 func makeResolveServiceNameConfig() *TransformProcessor {
 	return &TransformProcessor{
-		ErrorMode: "ignore",
-		MetricStatements: []TransformProcessorMetricStatements{
-			{
-				Context: "resource",
-				Statements: []string{
-					"set(attributes[\"service.name\"], attributes[\"kyma.kubernetes_io_app_name\"]) where attributes[\"service.name\"] == nil",
-					"set(attributes[\"service.name\"], attributes[\"kyma.app_name\"]) where attributes[\"service.name\"] == nil",
-					"set(attributes[\"service.name\"], attributes[\"k8s.deployment.name\"]) where attributes[\"service.name\"] == nil",
-					"set(attributes[\"service.name\"], attributes[\"k8s.daemonset.name\"]) where attributes[\"service.name\"] == nil",
-					"set(attributes[\"service.name\"], attributes[\"k8s.statefulset.name\"]) where attributes[\"service.name\"] == nil",
-					"set(attributes[\"service.name\"], attributes[\"k8s.job.name\"]) where attributes[\"service.name\"] == nil",
-					"set(attributes[\"service.name\"], attributes[\"k8s.pod.name\"]) where attributes[\"service.name\"] == nil",
-					"set(attributes[\"service.name\"], \"unknown_service\") where attributes[\"service.name\"] == nil",
-				},
-			},
-		},
+		ErrorMode:        "ignore",
+		MetricStatements: gatewayprocs.MakeResolveServiceNameStatements(),
 	}
 }
