+++ internal/otelcollector/config/trace/gateway/processors.go
@@ -2,47 +2,10 @@ package gateway
 
 import (
 	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config"
-	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config/servicename"
+	"github.com/kyma-project/telemetry-manager/internal/otelcollector/config/gatewayprocs"
 )
 
 func makeProcessorsConfig() Processors {
-	k8sAttributes := []string{
-		"k8s.pod.name",
-		"k8s.node.name",
-		"k8s.namespace.name",
-		"k8s.deployment.name",
-		"k8s.statefulset.name",
-		"k8s.daemonset.name",
-		"k8s.cronjob.name",
-		"k8s.job.name",
-	}
-
-	podAssociations := []config.PodAssociations{
-		{
-			Sources: []config.PodAssociation{
-				{
-					From: "resource_attribute",
-					Name: "k8s.pod.ip",
-				},
-			},
-		},
-		{
-			Sources: []config.PodAssociation{
-				{
-					From: "resource_attribute",
-					Name: "k8s.pod.uid",
-				},
-			},
-		},
-		{
-			Sources: []config.PodAssociation{
-				{
-					From: "connection",
-				},
-			},
-		},
-	}
-
 	return Processors{
 		BaseProcessors: config.BaseProcessors{
 			Batch: &config.BatchProcessor{
@@ -55,15 +18,7 @@ func makeProcessorsConfig() Processors {
 				LimitPercentage:      60,
 				SpikeLimitPercentage: 40,
 			},
-			K8sAttributes: &config.K8sAttributesProcessor{
-				AuthType:    "serviceAccount",
-				Passthrough: false,
-				Extract: config.ExtractK8sMetadata{
-					Metadata: k8sAttributes,
-					Labels:   servicename.ExtractLabels(),
-				},
-				PodAssociation: podAssociations,
-			},
+			K8sAttributes: gatewayprocs.MakeK8sAttributesProcessorConfig(),
 			Resource: &config.ResourceProcessor{
 				Attributes: []config.AttributeAction{
 					{
@@ -80,5 +35,12 @@ func makeProcessorsConfig() Processors {
 			},
 		},
 		ResolveServiceName: makeResolveServiceNameConfig(),
+	}
+}
+
+func makeResolveServiceNameConfig() *TransformProcessor {
+	return &TransformProcessor{
+		ErrorMode:       "ignore",
+		TraceStatements: gatewayprocs.MakeResolveServiceNameStatements(),
 	}
 }
